requirements.txt
torch==2.0.1
transformers==4.30.2
pillow==9.5.0
requests==2.31.0
tqdm==4.65.0
numpy==1.24.3
sentencepiece==0.1.99
protobuf==3.20.3
einops==0.6.1
accelerate==0.20.3
bitsandbytes==0.41.1
marshmallow==3.19.0
gevent==21.12.0
openfabric_pysdk
CTransformers
marshmallow-enum
flask-apispec==0.11.4
